Moved client uis from subfolders to parent
