+++ src/client/components/ServiceConfigFlowPDF/ServiceConfigFlow.js
@@ -1,9 +1,11 @@
 import React from 'react';
 import { Button, Nav, NavItem, Tab, Row } from 'react-bootstrap';
 import ajax from 'superagent-bluebird-promise';
-import ServiceConfigFlow1 from '../ServiceConfigFlow1'
-import ServiceConfigFlow2 from '../../common/ServiceConfigFlowTaCCustomer'
-import ServiceConfigFlow3 from '../ServiceConfigFlow3'
+
+import ServiceConfigFlow1 from './ServiceConfigFlow1.js'
+import ServiceConfigFlow2 from '../common/ServiceConfigFlowTaCCustomer'
+import ServiceConfigFlow3 from './ServiceConfigFlow3.js'
+import ServiceConfigFlow4 from './ServiceConfigFlow4.js'
 
 // A workaround to prevent a browser warning about unknown properties 'active', 'activeKey' and 'activeHref'
 // in DIV element.
@@ -13,18 +15,19 @@ const MyDiv = () =>
 };
 
 
-
 export default class ServiceConfigFlow extends React.Component
 {
     static propTypes = {
         currentTab : React.PropTypes.number,
         lastValidTab : React.PropTypes.number,
+        cancelWorkflow : React.PropTypes.func,
         inputType: React.PropTypes.string
     };
 
     static defaultProps = {
         currentTab : 1,
         lastValidTab : 1,
+        cancelWorkflow : () => { alert('Canceled!'); },
         inputType: null
     };
 
@@ -35,13 +38,23 @@ export default class ServiceConfigFlow extends React.Component
         this.state = {
             currentTab : this.props.currentTab,
             lastValidTab : this.props.lastValidTab,
+            cancelWorkflow : this.props.cancelWorkflow,
             inputType: this.props.inputType
         };
+
+        this.getInChannelConfig().catch(e => this.addInChannelConfig());
     }
 
-    ///////////////////////////////////////////////
-    // Webservice calls
-    ///////////////////////////////////////////////
+
+    static contextTypes = {
+        i18n : React.PropTypes.object.isRequired,
+    };
+
+
+
+    ///////////////////////////////////////////////////
+    // Helper methods to fetch or post data from/to server
+    ///////////////////////////////////////////////////
 
     //
     // InChannelConfig
@@ -55,30 +68,28 @@ export default class ServiceConfigFlow extends React.Component
 
     addInChannelConfig = () =>
     {
-console.log("++ addInChannelConfig -> paper/new");
         return ajax.post('/einvoice-send/api/config/inchannel')  // !!! /current
             .set('Content-Type', 'application/json')
             .send({
-                inputType : 'paper',
+                inputType : 'pdf',
                 status: 'new'
             })
             .promise();
     }
 
     updateInChannelConfig = (values) => {
-console.log("++ updateInChannelConfig -> values: ", values);
         return ajax.put('/einvoice-send/api/config/inchannel')  // !!! /current
             .set('Content-Type', 'application/json')
-            .send(values).promise();
+            .send(values)
+            .promise();
     }
 
     //
     // InChannelContract
     //
     getInChannelContract = (customerId) => {
-console.log("++ getInChannelContract -> customerId: ", customerId);
         return ajax.get('/einvoice-send/api/config/inchannelcontract/' + customerId)
-            .set('Content-Type', 'application/json')   // ??? really needed?
+            .set('Content-Type', 'application/json')
             .promise();
     }
 
@@ -89,7 +100,7 @@ console.log("++ addInChannelContract -> customerId =" + customerId + ", status =
             .set('Content-Type', 'application/json')
             .send({
                 customerId : customerId,
-                inputType : 'paper',
+                inputType : 'pdf',
                 status : status
             })
             .promise();
@@ -101,18 +112,18 @@ console.log("++ updateInChannelContract -> customerId / status: ", customerId, s
             .set('Content-Type', 'application/json')
             .send({
                 customerId : customerId,
-                inputType : 'paper',
+                inputType : 'pdf',
                 status : status
             })
             .promise();
     }
 
 
-    ////////////////////////////////////////
+    /////////////////////////////////////////////////////////
     // Events
-    ////////////////////////////////////////
+    /////////////////////////////////////////////////////////
     setApprovedOcTc = () => {
-        return this.updateInChannelConfig({'status':'approvedTC'});
+        return this.updateInChannelConfig({'status':'approved'});
     }
 
     setApprovedCustomerTc = () => {
@@ -144,13 +155,20 @@ console.log("InchannelConfig found: ", data);
         })
     }
 
+
     finalApprove = () => {
-        return ajax.get('/api/config/inchannel/approved')
+console.log(" ---- 1. finalApprove");
+        return ajax.put('/einvoice-send/api/config/finish')
             .promise()
         .then(() => {
+console.log(" ---- 2. finalApprove");
             this.props.finalizeFlow();
-        });
+        })
+        .catch((e) => {
+            alert ("The forwarding to the Invoice mapping team did not succeed. Please retry.")
+        })
     }
+
 
     setCurrentTab = (tabNo) => {
         this.setState({ currentTab : tabNo });
@@ -168,7 +186,6 @@ console.log("InchannelConfig found: ", data);
 
 
 
-
     render()
     {
         var visible = {
@@ -182,11 +199,11 @@ console.log("InchannelConfig found: ", data);
                         <div className="container">
                             <section className="header">
                                 <h1>
-                                    Service Configuration Flow
+                                    {this.context.i18n.getMessage('ServiceConfigFlow.header')}
                                     <div className="control-bar text-right pull-right">
                                         <Button onClick={ () => this.props.gotoStart()}>
                                             <i className="fa fa-angle-left"/>
-                                            &nbsp;&nbsp;Back to Type Selection
+                                            &nbsp;&nbsp;{this.context.i18n.getMessage('ServiceConfigFlow.backToTypeSelection')}
                                         </Button>
                                     </div>
                                 </h1>
@@ -225,6 +242,7 @@ console.log("InchannelConfig found: ", data);
 }
 
 
+
 class PaperNav extends React.Component {
     render() {
         if (this.props.customerTermsAndConditions) {
@@ -239,6 +257,9 @@ class PaperNav extends React.Component {
                         <span className="round-tab"><i className="glyphicon glyphicon-pencil"/></span>
                     </NavItem>
                     <NavItem eventKey={3} disabled={ this.props.currentTab < 3 }>
+                        <span className="round-tab"><i className="glyphicon glyphicon-search"/></span>
+                    </NavItem>
+                    <NavItem eventKey={4} disabled={ this.props.currentTab < 4 }>
                         <span className="round-tab"><i className="glyphicon glyphicon-ok"/></span>
                     </NavItem>
                 </Nav>
@@ -253,6 +274,9 @@ class PaperNav extends React.Component {
                         <span className="round-tab"><i className="glyphicon glyphicon-pencil"/></span>
                     </NavItem>
                     <NavItem eventKey={2} disabled={ this.props.currentTab < 2 }>
+                        <span className="round-tab"><i className="glyphicon glyphicon-search"/></span>
+                    </NavItem>
+                    <NavItem eventKey={3} disabled={ this.props.currentTab < 3 }>
                         <span className="round-tab"><i className="glyphicon glyphicon-ok"/></span>
                     </NavItem>
                 </Nav>
@@ -281,8 +305,14 @@ class PaperTabContent extends React.Component {
                     </Tab.Pane>
                     <Tab.Pane eventKey={3}>
                         <ServiceConfigFlow3
-                            onNext={ () => { this.props.finalApprove() } }
+                            onNext={ () => { this.props.setCurrentTab(4) } }
                             onPrevious={ () => this.props.setCurrentTab(2) }
+                            voucher = {this.props.voucher}/>
+                    </Tab.Pane>
+                    <Tab.Pane eventKey={4}>
+                        <ServiceConfigFlow4
+                            onNext={ () => { this.props.finalApprove() } }
+                            onPrevious={ () => this.props.setCurrentTab(3) }
                             voucher = {this.props.voucher}/>
                     </Tab.Pane>
                 </Tab.Content>
@@ -298,8 +328,14 @@ class PaperTabContent extends React.Component {
                     </Tab.Pane>
                     <Tab.Pane eventKey={2}>
                         <ServiceConfigFlow3
-                            onNext={ () => { this.props.finalApprove() } }
+                            onNext={ () => { this.props.setCurrentTab(3) } }
                             onPrevious={ () => this.props.setCurrentTab(1) }
+                            voucher = {this.props.voucher}/>
+                    </Tab.Pane>
+                    <Tab.Pane eventKey={3}>
+                        <ServiceConfigFlow4
+                            onNext={ () => { this.props.finalApprove() } }
+                            onPrevious={ () => this.setCurrentTab(2) }
                             voucher = {this.props.voucher}/>
                     </Tab.Pane>
                 </Tab.Content>
