--- src/client/components/ServiceConfigFlowPDF/ServiceConfigFlow3/i18n/en.js
@@ -1,2 +1,0 @@
-export default {
-}
