--- src/client/components/ServiceConfigFlowPDF/ServiceConfigFlow4/i18n/en.js
@@ -1,2 +1,0 @@
-export default {
-}
