--- src/client/components/ServiceConfigFlowPDF/ServiceConfigFlow4/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "ServiceConfigFlow4",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./ServiceConfigFlow4.js"
-}
