--- src/client/components/ServiceConfigFlowPDF/ServiceConfigFlowFrame/i18n/de.js
@@ -1,10 +1,0 @@
-export default {
-    ServiceConfigFlow : {
-        test:"Hallo",
-        one:"eins"
-    },
-    ServiceConfigFlow2 : {
-        test:"Hallo",
-        one:"eins"
-    }
-}
