--- src/client/components/ServiceConfigFlowPDF/ServiceConfigFlowFrame/i18n/en.js
@@ -1,12 +1,0 @@
-export default {
-    ServiceConfigFlow : {
-        info: "part of pdf/ServiceConfigFlowFrame",
-        test:"Hello",
-        one: "en-One"
-    },
-    ServiceConfigFlow2 : {
-        info: "part of pdf/ServiceConfigFlowFrame",
-        test:"Hello",
-        one: "en-One"
-    }
-}
