--- src/client/components/ServiceConfigFlowPaper/ServiceConfigFlow3/i18n/locales.js
@@ -1,12 +1,0 @@
-import en from './en';
-import de from './de';
-export default [
-  {
-    locales: ['en'],
-    messages: en
-  },
-  {
-    locales: ['de'],
-    messages: de
-  }
-];
