--- src/client/components/ServiceConfigFlowPaper/ServiceConfigFlow3/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "ServiceConfigFlow3",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./ServiceConfigFlow3.js"
-}
