--- src/client/components/ServiceConfigFlowPaper/ServiceConfigFlowFrame/i18n/de.js
@@ -1,2 +1,0 @@
-export default {
-}
