--- src/client/components/ServiceConfigFlowPaper/ServiceConfigFlowFrame/i18n/en.js
@@ -1,2 +1,0 @@
-export default {
-}
