--- src/client/components/ServiceConfigFlowPaper/ServiceConfigFlowFrame/i18n0/index.js
@@ -1,12 +1,0 @@
-import en from './en';
-import de from './de';
-export default [
-  {
-    locales: ['en'],
-    messages: en
-  },
-  {
-    locales: ['de'],
-    messages: de
-  }
-];
