--- src/client/components/ServiceConfigFlowStart/i18n/de.js
@@ -1,2 +1,0 @@
-export default {
-}
