+++ src/server/db/models/voucher.js
@@ -27,6 +27,7 @@ module.exports.init = function(db, config)
             type : DataTypes.STRING(30),
             allowNull : true
         },
+        // new, approved
         status : {
             type : DataTypes.STRING(100),
             allowNull : false,
